+++ src/dataset/domain/repositories/DatasetRepository.ts
@@ -8,6 +8,7 @@ import { DatasetVersionSummaryInfo } from '../models/DatasetVersionSummaryInfo'
 import { DatasetDeaccessionDTO } from '../useCases/DTOs/DatasetDTO'
 import { DatasetDownloadCount } from '../models/DatasetDownloadCount'
 import { FormattedCitation, CitationFormat } from '../models/DatasetCitation'
+import { DatasetTemplate } from '../models/DatasetTemplate'
 
 export interface DatasetRepository {
   getByPersistentId: (
@@ -54,4 +55,5 @@ export interface DatasetRepository {
     version: string,
     format: CitationFormat
   ) => Promise<FormattedCitation>
+  getTemplates: (collectionIdOrAlias: number | string) => Promise<DatasetTemplate[]>
 }