+++ src/stories/file/File.stories.tsx
@@ -7,6 +7,11 @@ import { FileMockLoadingRepository } from './FileMockLoadingRepository'
 import { FileMockNoDataRepository } from './FileMockNoDataRepository'
 import { FileMother } from '../../../tests/component/files/domain/models/FileMother'
 import { DatasetMockRepository } from '../dataset/DatasetMockRepository'
+import { ExternalToolsProvider } from '@/shared/contexts/external-tools/ExternalToolsProvider'
+import { ExternalToolsMockRepository } from '../shared-mock-repositories/externalTools/ExternalToolsMockRepository'
+import { FakerHelper } from '@tests/component/shared/FakerHelper'
+import { ExternalToolsMother } from '@tests/component/externalTools/domain/models/ExternalToolsMother'
+import { DataverseInfoMockRepository } from '../shared-mock-repositories/info/DataverseInfoMockRepository'
 
 const meta: Meta<typeof File> = {
   title: 'Pages/File',
@@ -26,6 +31,7 @@ export const Default: Story = {
     <File
       repository={new FileMockRepository()}
       datasetRepository={new DatasetMockRepository()}
+      dataverseInfoRepository={new DataverseInfoMockRepository()}
       id={56}
     />
   )
@@ -36,6 +42,7 @@ export const Restricted: Story = {
     <File
       repository={new FileMockRepository(FileMother.createRestricted())}
       datasetRepository={new DatasetMockRepository()}
+      dataverseInfoRepository={new DataverseInfoMockRepository()}
       id={56}
     />
   )
@@ -46,6 +53,7 @@ export const RestrictedWithAccessGranted: Story = {
     <File
       repository={new FileMockRepository(FileMother.createRestrictedWithAccessGranted())}
       datasetRepository={new DatasetMockRepository()}
+      dataverseInfoRepository={new DataverseInfoMockRepository()}
       id={56}
     />
   )
@@ -56,6 +64,7 @@ export const Loading: Story = {
     <File
       repository={new FileMockLoadingRepository()}
       datasetRepository={new DatasetMockRepository()}
+      dataverseInfoRepository={new DataverseInfoMockRepository()}
       id={56}
     />
   )
@@ -66,7 +75,68 @@ export const FileNotFound: Story = {
     <File
       repository={new FileMockNoDataRepository()}
       datasetRepository={new DatasetMockRepository()}
+      dataverseInfoRepository={new DataverseInfoMockRepository()}
       id={56}
     />
+  )
+}
+
+export const WithMultipleExternalTools: Story = {
+  render: () => (
+    <ExternalToolsProvider externalToolsRepository={new ExternalToolsMockRepository()}>
+      <File
+        repository={new FileMockRepository(FileMother.createRealistic())}
+        datasetRepository={new DatasetMockRepository()}
+        dataverseInfoRepository={new DataverseInfoMockRepository()}
+        id={56}
+        toolTypeSelectedQueryParam="preview"
+      />
+    </ExternalToolsProvider>
+  )
+}
+
+const externalToolsRepositoryOnlyPreviewTool = new ExternalToolsMockRepository()
+externalToolsRepositoryOnlyPreviewTool.getExternalTools = () => {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve([ExternalToolsMother.createFilePreviewTool()])
+    }, FakerHelper.loadingTimout())
+  })
+}
+
+export const WithOnlyOnePreviewExternalTool: Story = {
+  render: () => (
+    <ExternalToolsProvider externalToolsRepository={externalToolsRepositoryOnlyPreviewTool}>
+      <File
+        repository={new FileMockRepository(FileMother.createRealistic())}
+        datasetRepository={new DatasetMockRepository()}
+        dataverseInfoRepository={new DataverseInfoMockRepository()}
+        id={56}
+        toolTypeSelectedQueryParam="preview"
+      />
+    </ExternalToolsProvider>
+  )
+}
+
+const externalToolsRepositoryOnlyQueryTool = new ExternalToolsMockRepository()
+externalToolsRepositoryOnlyQueryTool.getExternalTools = () => {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve([ExternalToolsMother.createFileQueryTool()])
+    }, FakerHelper.loadingTimout())
+  })
+}
+
+export const WithOnlyOneQueryExternalTool: Story = {
+  render: () => (
+    <ExternalToolsProvider externalToolsRepository={externalToolsRepositoryOnlyQueryTool}>
+      <File
+        repository={new FileMockRepository(FileMother.createRealistic())}
+        datasetRepository={new DatasetMockRepository()}
+        dataverseInfoRepository={new DataverseInfoMockRepository()}
+        id={56}
+        toolTypeSelectedQueryParam="query"
+      />
+    </ExternalToolsProvider>
   )
 }